requirements.txt
pytesseract==0.3.7
matplotlib==3.3.2
kraken==2.0.8
nltk==3.5
spacy==3.0.3
numpy==1.19.3
opencv_python==4.5.1.48
pandas==1.2.1
Pillow==8.1.0
